+++ lib/previews_guest.dart
@@ -13,7 +13,7 @@ library;
 export 'src/ui_catalog/guest.dart' show CatalogGuest, CatalogKnobs;
 export 'src/ui_catalog/axes.dart' show CatalogAxes;
 export 'src/ui_catalog/entries.dart' show CatalogEntries;
-export 'src/ui_catalog/clock.dart' show previewClockOrigin, withPreviewClock;
+export 'src/ui_catalog/clock.dart' show pinnedClockOrigin, withPreviewClock;
 export 'src/ui_catalog/guest_keyboard.dart' show GuestKeyboard;
 export 'src/ui_catalog/guest_text_input.dart' show GuestTextInput;
 export 'src/ui_catalog/catalog_keyboard.dart'